// packages/weapp-tailwindcss/src/bundlers/vite/generate-bundle/types.ts
export interface RememberedCssSource {
  outputFile: string
  rawSource: string
  sourceFile: string
}